Refactor Get Logic to fix idempotency issue.
Take for example the following records:
- `a.b.c.d`
- `b.c.d`

When using `Get-DnsServerResourceRecord`
to retrieve record `b.c.d` two records are returned.
This causes powershell to return a list of two values back to go
and results in record `b.c.d` needing to be removed and recreated.

This change adds logic to filter out the additional records
based on the record name being retrieved.